Update exercise "15.01. ZooShop". Add dog class and some tests.

+++ DevStyleR/C++ Programing Part 1/15.01. ZooShop/Dog.h
@@ -4,7 +4,7 @@
 
 class Dog : public Animal {
 public:
-	Dog(string, string, int, string);
+	Dog(string, string, double, string);
 
 	void setBark(string);
